+++ lib/src/chart/line_chart/line_chart.dart
@@ -97,7 +97,8 @@ class _LineChartState extends AnimatedWidgetBaseState<LineChart> {
         }
 
         _showingTouchedTooltips.clear();
-        _showingTouchedTooltips.add(ShowingTooltipIndicators(0, sortedLineSpots));
+        _showingTouchedTooltips
+            .add(ShowingTooltipIndicators(0, sortedLineSpots.map((e) => e.spotIndex).toList()));
       });
     } else {
       setState(() {